+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 Changelog
 =========
 
+v0.4.7
+------
+
+* Bug fixes
+
+v0.4.6
+------
+
+* Added navigation option
+* Made left menu scrollable in lagoon template
+* Added livereload property
+* Bug fixes
+
 v0.4.5
 ------